moves hardcoded business to list component and allows for business input to be an object. Cleans BusinessList code to use array of li's and removes unused code
