+++ src/components/BusinessList.js
@@ -1,16 +1,34 @@
-//import './BusinessList.css';
-import Business from "./Business"
+import Business from "./Business";
+
+const hardcodedBusiness = {
+    name: "Uncle Chip's",
+    address: "1100 North Capitol St",
+    city: "Washington",
+    state: "DC",
+    zipcode: "20001",
+    category: "American",
+    rating: "4.8",
+    review_count: "237",
+    src: "https://lh3.googleusercontent.com/p/AF1QipOQI5WU8mTmYDxDXuySRrOlI05XGWM42wnk6X0B=s1360-w1360-h1020",
+    alt: "Uncle Chip's businesss photo"
+}; //DEBUG code, used for testing
+
+let businessArray = [
+    <Business business={hardcodedBusiness}/>,
+    <Business business={hardcodedBusiness}/>,
+    <Business business={hardcodedBusiness}/>,
+    <Business business={hardcodedBusiness}/>,
+    <Business business={hardcodedBusiness}/>,
+    <Business business={hardcodedBusiness}/>
+];
 
 function BusinessList() {
-  return (
+    const list = businessArray.map(business=><li>{business}</li>);
+
+    return (
     <div className="BusinessList">
       <ul>
-        <li><Business /></li>
-        <li><Business /></li>
-        <li><Business /></li>
-        <li><Business /></li>
-        <li><Business /></li>
-        <li><Business /></li>
+        {list}
       </ul>
     </div>
   );
